Fix 18 CodeQL code scanning alerts
- Fix potential use-after-free: reset videoBuffers/nVideoBuffers after
  free() in CloseAdsd3500() and SetFrameType() (adsd3500_util.cpp)
- Fix multiplication result converted to larger type by widening one
  operand to size_t before multiplying in ADIView.cpp, ADIView_neon.cpp,
  stb_image.h, and CCB_CRC_READ.cpp
- Add explicit contents: write permission to
  automerge_release_to_dev.yml workflow

+++ tools/nvm_tools/CCB_CRC_READ/CCB_CRC_READ.cpp
@@ -172,7 +172,7 @@ memset(data, 0x00, CTRL_SIZE);
 	{
 		packetsNeeded = packetsNeeded+1;
 	}
-        copy_buffer = (uint8_t*)malloc(packetsNeeded * chunkSize);
+        copy_buffer = (uint8_t*)malloc(static_cast<size_t>(packetsNeeded) * chunkSize);
 	std::cout<<"Size of Binary :: "<<sizeOfBinary<<std::endl;
 	std::cout<<"ChunkSize :: "<<chunkSize<<std::endl;
 	std::cout<<"Packets Needed: "<<packetsNeeded<<std::endl;
